deck
deck is a new layout for DWM (inspired by the TTWM window manager). It
is like tile(), but with monocle in the stack. You have a master client
(complete with nmaster et all), but the stack clients now overlap one
another as in monocle(). (The stacked clients are like a deck of cards,
hence the name).

Tile :
+-----------------+--------+
|                 |        |
|                 |  S1    |
|                 |        |
|        M        +--------+
|                 |        |
|                 |   S2   |
|                 |        |
+-----------------+--------+

Deck :
+-----------------+--------+
|                 |        |
|                 |        |
|                 |        |
|        M        |   S1   |
|                 |        |
|                 |        |
|                 |        |
+-----------------+--------+

URL: http://dwm.suckless.org/patches/deck